requirements.txt
Flask
flask-cors
PyPDF2
python-docx
wordcloud
matplotlib
gunicorn
